+++ tests/integration/components/bs-form/element-test.js
@@ -23,6 +23,7 @@ import hbs from 'htmlbars-inline-precompile';
 import setupNoDeprecations from '../../../helpers/setup-no-deprecations';
 import { gte } from 'ember-compatibility-helpers';
 import sinon from 'sinon';
+import { tracked } from '@glimmer/tracking';
 
 const formLayouts = ['vertical', 'horizontal', 'inline'];
 
@@ -391,28 +392,48 @@ module('Integration | Component | bs-form/element', function (hooks) {
   });
 
   test('Custom controls are supported', async function (assert) {
-    this.set(
-      'model',
-      EmberObject.create({
-        gender: 'male',
-      })
-    );
+    let model = new (class {
+      @tracked gender = 'male';
+    })();
+
+    this.set('model', model);
     await render(hbs`
       <BsForm @model={{model}} as |form|>
-        <form.element @label="Gender" @property="gender" @showAllValidations={{true}} @hasValidator={{true}} as |el|>
-          <div id="value">{{el.value}}</div>
-          <div id="id">{{el.id}}</div>
-          <div id="validation">{{el.validation}}</div>
-          <div id="control">{{el.control}}</div>
+        <form.element
+          @controlType="textarea"
+          @label="Gender"
+          @property="gender"
+          @showAllValidations={{true}}
+          @hasValidator={{true}}
+          as |el|
+        >
+          <input
+            type="text"
+            id={{el.id}}
+            value={{el.value}}
+            class={{el.validation}}
+            {{on "input" (action el.setValue value="target.value")}}
+          >
+          <el.control />
         </form.element>
       </BsForm>
     `);
 
-    assert.dom('#value').exists({ count: 1 }, 'block template is rendered');
-    assert.dom('#value').hasText('male', 'value is yielded to block template');
-    assert.dom('#id').hasAnyText('id is yielded to block template');
-    assert.dom('#validation').hasText('success');
-    assert.dom('#control input[type=text]').exists({ count: 1 }, 'control component is yielded');
+    assert.dom('input').exists({ count: 1 }, 'block template is rendered');
+    assert.dom('input').hasValue('male', 'value is yielded to block template');
+    assert.dom('input').hasAttribute('id', { any: true }, 'id is yielded to block template');
+    assert.dom('input').hasClass('success', 'validation status is yielded to block template');
+    assert.dom('textarea').exists({ count: 1 }, 'control component is yielded');
+
+    await fillIn('input', 'female');
+    assert.equal(model.gender, 'female', 'yielded setValue function updates property of model');
+    assert.dom('input').hasValue('female', 'yielded value is updated with setValue function');
+    assert.dom('textarea').hasValue('female', 'value of yielded control component is updated');
+
+    await fillIn('textarea', 'diverse');
+    assert.equal(model.gender, 'diverse', 'yielded control component updates property of model');
+    assert.dom('input').hasValue('diverse', 'yielded value is updated if changed through control component');
+    assert.dom('textarea').hasValue('diverse', 'value of yielded control component is updated');
   });
 
   test('required property propagates', async function (assert) {